Move config-realted stuff into a submodule by it's own.
